// SqGoods/ClientApp/src/api-client.generated.ts
export interface FilterBoolModel {
}

export interface FilterAndModel extends FilterBoolModel {
    items: FilterBoolModel[];
    type: string;
}

export interface FilterOrModel extends FilterBoolModel {
    items: FilterBoolModel[];
    type: string;
}